+++ examples/glwindow.rs
@@ -3,7 +3,8 @@ use fltk::{app::*, gl::*, window::*};
 pub fn main() {
     let app = App::default().set_scheme(AppScheme::Gleam);
     let mut wind = GlWindow::new(100, 100, 800, 600, "GlWindow Example");
-
+    wind.set_mode(Mode::Opengl3);
+    
     wind.end();
     wind.show();