+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 ### Changes
-- Translate gl functions available via fltk.
+- Translate gl functions exposed via fltk.
+- Add mode and set_mode to GlWindow.
 
 ## [0.4.11] - 2020-05-23
 ### Changes